requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
gspread>=5.10.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
Pillow>=10.0.0
python-dateutil>=2.8.0